+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.EjercicioDeProgramaciónOrientadaAObjetosPOO"
         tools:targetApi="31">
         <activity
-            android:name=".AddActivity"
+            android:name=".Activities.AddActivity"
             android:exported="false" />
         <activity
-            android:name=".SearchActivity"
+            android:name=".Activities.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".DeleteActivity"
+            android:name=".Activities.DeleteActivity"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name=".Activities.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
