+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Activities/DeleteActivity.kt
@@ -1,16 +1,21 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Activities
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.ejerciciodeprogramacinorientadaaobjetospoo.R
+import com.example.ejerciciodeprogramacinorientadaaobjetospoo.databinding.ActivityDeleteBinding
 
-class SearchActivity : AppCompatActivity() {
+class DeleteActivity : AppCompatActivity() {
+    lateinit var binding : ActivityDeleteBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        binding = ActivityDeleteBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_search)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
