+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Activities/SearchActivity.kt
@@ -1,20 +1,17 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Activities
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.example.ejerciciodeprogramacinorientadaaobjetospoo.databinding.ActivityDeleteBinding
+import com.example.ejerciciodeprogramacinorientadaaobjetospoo.R
 
-class DeleteActivity : AppCompatActivity() {
-    lateinit var binding : ActivityDeleteBinding
-
+class SearchActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityDeleteBinding.inflate(layoutInflater)
-        setContentView(binding.root)
         enableEdgeToEdge()
+        setContentView(R.layout.activity_search)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
