+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Attributes/AttributesSmart.kt
@@ -1,0 +1,9 @@
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Attributes
+
+import com.example.ejerciciodeprogramacinorientadaaobjetospoo.Classes.Sim
+
+data class AttributesSmart(
+     val screenSize: String,
+     val operativeSystem: String,
+     val sim: Sim
+)
