--- app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/AttributesSmart.kt
@@ -1,7 +1,0 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
-
- data class AttributesSmart(
-     val screenSize: String,
-     val operativeSystem: String,
-     val sim: Sim
-)
