+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Classes/Sim.kt
@@ -1,0 +1,5 @@
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Classes
+
+enum class Sim {
+    SINGLE, DUAL
+}
