+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Classes/Smartphone.kt
@@ -1,4 +1,4 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Classes
 
 open class Smartphone(brand : String, model: String, state : State, id : String) :
     Devices(brand, model, state, id)  {
