+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Classes/State.kt
@@ -1,0 +1,5 @@
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Classes
+
+enum class State {
+    NEW, USED, REPAIRING, BROKEN
+}
