+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Classes/Tablet.kt
@@ -1,4 +1,4 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Classes
 
 class Tablet(brand: String, model: String, state: State, id: String) :
     Devices(brand, model, state, id) {
