--- app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/DevicesRegister.kt
@@ -1,4 +1,0 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
- object DevicesRegister {
-     val listOfDevices = mutableListOf<Devices>()
- }
