+++ app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Object/DevicesRegister.kt
@@ -1,0 +1,7 @@
+package com.example.ejerciciodeprogramacinorientadaaobjetospoo.Object
+
+import com.example.ejerciciodeprogramacinorientadaaobjetospoo.Classes.Devices
+
+object DevicesRegister {
+     val listOfDevices = mutableListOf<Devices>()
+ }
