--- app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/Sim.kt
@@ -1,5 +1,0 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
-
-enum class Sim {
-    SINGLE, DUAL
-}
