--- app/src/main/java/com/example/ejerciciodeprogramacinorientadaaobjetospoo/State.kt
@@ -1,5 +1,0 @@
-package com.example.ejerciciodeprogramacinorientadaaobjetospoo
-
-enum class State {
-    NEW, USED, REPAIRING, BROKEN
-}
