+++ app/src/main/res/layout/activity_add.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddActivity">
+    tools:context=".Activities.AddActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
