+++ app/src/main/res/layout/activity_delete.xml
@@ -7,44 +7,43 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".DeleteActivity">
+    tools:context=".Activities.DeleteActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="80dp"
         android:layout_marginTop="0dp"
         android:background="#E91E63"
-        android:orientation="vertical"
-        android:id="@+id/pepe">
+        android:orientation="vertical">
 
         <TextView
+            android:id="@+id/text_delete_activity"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:gravity="center"
-            android:text="Borrar"
+            android:text="@string/delete_message"
             android:textColor="@color/black"
-            android:textSize="30sp"
-            android:id="@+id/qsymaster" />
+            android:textSize="30sp" />
 
     </LinearLayout>
 
     <TextView
+        android:id="@+id/text_input_ID_delete"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="120dp"
         android:gravity="center"
-        android:text="Ingrese el ID del celular a eliminar: "
+        android:text="@string/input_id_delete"
         android:textSize="20sp"
-        android:textStyle="bold"
-        android:id="@+id/asdasd" />
+        android:textStyle="bold" />
 
     <LinearLayout
-        android:id="@+id/input_card_delete"
+        android:id="@+id/input_delete"
         android:layout_width="match_parent"
         android:layout_height="40dp"
-        android:layout_marginStart="40dp"
+        android:layout_marginStart="30dp"
         android:layout_marginTop="2dp"
-        android:layout_marginEnd="40dp"
+        android:layout_marginEnd="30dp"
         android:background="@drawable/background_black">
 
         <EditText
@@ -53,8 +52,8 @@
             android:layout_margin="1dp"
             android:background="@drawable/input_background"
             android:hint="ID:"
-            android:paddingStart="10dp"
-            android:inputType="text">
+            android:inputType="text"
+            android:paddingStart="10dp">
 
         </EditText>
     </LinearLayout>
@@ -62,64 +61,51 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="300dp"
-        android:background="@color/white"
-        android:layout_marginTop="80dp"
         android:layout_marginStart="40dp"
-        android:layout_marginBottom="40dp"
+        android:layout_marginTop="80dp"
         android:layout_marginEnd="40dp"
-        android:orientation="vertical"
+        android:layout_marginBottom="40dp"
+        android:background="@color/white"
         android:gravity="center"
-        >
+        android:orientation="vertical">
+
+
 
         <TextView
+            android:id="@+id/show_id"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:maxLines="1"
+            android:text="@string/id_response"
+            android:textSize="25sp" />
 
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/object_deleted"
-            android:text="Objeto eliminado:"
-            android:textSize="25dp"
-            android:gravity="center" />
         <TextView
+            android:id="@+id/show_state"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:gravity="center"
             android:maxLines="1"
-            android:id="@+id/show_name"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
+            android:text="@string/model_response"
+            android:textSize="25sp" />
+
         <TextView
+            android:id="@+id/show_model"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:gravity="center"
             android:maxLines="1"
-            android:id="@+id/show_x"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
+            android:text="@string/state_response"
+            android:textSize="25sp" />
+
         <TextView
+            android:id="@+id/show_brand"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:gravity="center"
             android:maxLines="1"
-            android:id="@+id/show_xx"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_yy"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_yyy"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
+            android:text="@string/brand_response"
+            android:textSize="25sp" />
 
     </LinearLayout>
