+++ app/src/main/res/layout/activity_search.xml
@@ -7,36 +7,36 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".SearchActivity">
+    tools:context=".Activities.SearchActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="80dp"
         android:layout_marginTop="0dp"
         android:background="@color/blue"
-        android:orientation="vertical"
-        android:id="@+id/pepe">
+        android:orientation="vertical">
 
         <TextView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:gravity="center"
-            android:text="Borrar"
+            android:text="Buscar"
             android:textColor="@color/black"
-            android:textSize="30sp"
-            android:id="@+id/qsymaster" />
+            android:textSize="30sp" />
 
     </LinearLayout>
 
     <TextView
+        android:id="@+id/delete_message"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="120dp"
+        android:layout_marginTop="10dp"
         android:gravity="center"
-        android:text="Ingrese el ID del celular a eliminar: "
+        android:text="@string/input_id_delete"
         android:textSize="20sp"
         android:textStyle="bold"
-        android:id="@+id/asdasd" />
+
+        />
 
     <LinearLayout
         android:id="@+id/input_card_delete"
@@ -48,78 +48,31 @@
         android:background="@drawable/background_black">
 
         <EditText
+            android:id="@+id/sv_search_activity"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_margin="1dp"
             android:background="@drawable/input_background"
-            android:hint="ID:"
-            android:paddingStart="10dp"
-            android:inputType="text">
+            android:hint="@string/search_by_id"
+            android:inputType="text"
+            android:paddingStart="15dp"
+            android:autofillHints=""
+            >
 
         </EditText>
     </LinearLayout>
 
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="300dp"
-        android:background="@color/white"
-        android:layout_marginTop="80dp"
+        android:layout_height="match_parent"
         android:layout_marginStart="40dp"
-        android:layout_marginBottom="40dp"
+        android:layout_marginTop="80dp"
         android:layout_marginEnd="40dp"
-        android:orientation="vertical"
+        android:layout_marginBottom="40dp"
+        android:background="@color/white"
         android:gravity="center"
-        >
+        android:orientation="vertical">
 
-        <TextView
-
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/object_deleted"
-            android:text="Objeto eliminado:"
-            android:textSize="25dp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_name"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_x"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_xx"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_yy"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:id="@+id/show_yyy"
-            android:text="Objeto eliminado:"
-            android:textSize="25sp"
-            android:gravity="center" />
 
     </LinearLayout>
